requirements.txt
openai==1.40.0
playwright==1.45.0
pydantic==2.7.4
python-dotenv==1.0.1
rich==13.7.1
pandas==2.2.2
lxml==5.2.2
httpx==0.27.0
